move utils.serialization
